+++ tests/geophires_x_tests/test_economics_sam.py
@@ -1137,14 +1137,3 @@ class EconomicsSamTestCase(BaseTestCase):
             m.Calculate()
 
         return m
-
-    def _handle_assert_logs_failure(self, ae: AssertionError):
-        if sys.version_info[:2] == (3, 8) and self._is_github_actions():
-            # FIXME - see
-            #  https://github.com/softwareengineerprogrammer/GEOPHIRES/actions/runs/19646240874/job/56262028512#step:5:344
-            _log.warning(
-                f'WARNING: Skipping logs assertion in GitHub Actions '
-                f'for Python {sys.version_info.major}.{sys.version_info.minor}'
-            )
-        else:
-            raise ae
